fix : kwh Double 변환 작업 완료

+++ src/main/java/live/ioteatime/apiservice/domain/MonthlyElectricity.java
@@ -22,7 +22,7 @@ public class MonthlyElectricity {
     @ManyToOne
     private Channel channel;
 
-    private Long kwh;
+    private Double kwh;
     private Long bill;
 
     @Embeddable
